few fixes + add count_lines_limit param

+++ change.bulkcleaner.php
@@ -68,7 +68,7 @@ function UsageAndExit($oP) {
 
     if ($bModeCLI) {
         $oP->p("USAGE:\n");
-        $oP->p("php change.bulkcleaner.php --auth_user=<login> --auth_pwd=<password> --bulk_size=<bulk_size> [--param_file=<file>]\n");
+        $oP->p("php change.bulkcleaner.php --auth_user=<login> --auth_pwd=<password> --bulk_size=<bulk_size> [-count_lines_limit=<count_lines_limit>] [--param_file=<file>]\n");
     } else {
         $oP->p("Optional parameters: verbose, param_file, status_only\n");
     }
@@ -84,11 +84,12 @@ function BulkDelete($iBulkSize)
 
     $sPrefix = \MetaModel::GetConfig()->Get('db_subname');
 
-    $aSqlQueries = [];
-
-    //$this->ProvisionDb($aSqlQueries, $sPrefix);
-
     $aIds= GetIds($sPrefix, $iBulkSize);
+
+    if (count($aIds) == 0){
+        return "No CMDBChange row to delete.";
+    }
+
     $sBulkDelete = sprintf("DELETE FROM %spriv_change WHERE id IN (%s)",
         $sPrefix,
         implode(',', $aIds)
@@ -96,7 +97,7 @@ function BulkDelete($iBulkSize)
 
     ExecuteQuery($sBulkDelete);
 
-    $sMsg= sprintf("%d ChangeOp row(s) deleted.", count($aIds));
+    $sMsg= sprintf("%d CMDBChange row(s) deleted.", count($aIds));
     IssueLog::Info($sMsg);
     return $sMsg;
 }
@@ -145,30 +146,69 @@ SQL;
     return $aIds;
 }
 
-function ProvisionDb($aSqlQueries, $sPrefix){
-    for ($i=0;$i<1000;$i++){
-        $aSqlQueries["INSERT$i"] = <<<SQL
-	INSERT INTO {$sPrefix}priv_change (`userinfo`, `date`, `origin`) 
-	values
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing'),
-	('test', NOW(), 'email-processing');
-SQL;
+/**
+ * @param $sPrefix
+ * @param int $iLimit
+ * @return int
+ * @throws CoreException
+ * @throws MySQLException
+ * @throws MySQLHasGoneAwayException
+ */
+function CountLinesToRemove($iLimit=0){
+    if ($iLimit==0){
+        return -1;
     }
 
+    $sPrefix = \MetaModel::GetConfig()->Get('db_subname');
+
+    $sSqlQuery = <<<SQL
+SELECT count(*) AS count FROM 
+(SELECT co.id FROM ${sPrefix}priv_change AS c 
+LEFT JOIN ${sPrefix}priv_changeop AS co 
+ON co.changeid = c.id 
+WHERE co.id IS NULL
+LIMIT
+) AS t;
+SQL;
+
+    if ($iLimit != -1){
+        $sSqlQuery = str_replace("LIMIT", " LIMIT {$iLimit} ", $sSqlQuery);
+    } else {
+        $sSqlQuery = str_replace("LIMIT", "", $sSqlQuery);
+    }
+
+    $oQueryResult = ExecuteQuery($sSqlQuery);
+
+    while($aRow = $oQueryResult->fetch_array()){
+        return (int) $aRow['count'];
+    }
+
+    return 0;
+}
+
+function ProvisionDb($iMax){
+    $sPrefix = \MetaModel::GetConfig()->Get('db_subname');
+
+    for ($i=0;$i<$iMax;$i++){
+        $sSqlQuery = <<<SQL
+	INSERT INTO {$sPrefix}priv_change (`userinfo`, `date`, `origin`) 
+	values
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing'),
+	("test{$i}", NOW(), 'email-processing');
+SQL;
+
+        ExecuteQuery($sSqlQuery);
+    }
+
+    return $iMax * 10;
 }
 
 if (!utils::IsModeCLI()){
@@ -224,6 +264,21 @@ try
         if ($oMutex->TryLock())
         {
             $iBulkDelete = ReadMandatoryParam($oP, 'bulk_size', 'integer');
+
+            $iCountLimit = utils::ReadParam('count_lines_limit', 0, true, 'integer');
+
+            //$iCreatedCount = ProvisionDb(100);
+            //return "ProvisionDb $iCreatedCount added lines";
+
+            $iCount = CountLinesToRemove($iCountLimit);
+            if ($iCount !== -1){
+                if ($iCount < $iCountLimit){
+                    $oP->p("Found exactly $iCount CMDBChange row(s) to delete");
+                }else{
+                    $oP->p("Found at least $iCount CMDBChange row(s) to delete");
+                }
+            }
+
             $oP->p(BulkDelete($iBulkDelete));
         }
         else
